[PM-19949] Provide ConfigDiskSource from `data` module
Make the `ConfigDiskSourceImpl` internal and provide `ConfigDiskSource` from the `data` module.
